+++ locale_generator/example/lib/simple_page_locale/simple_page_locale.dart
@@ -6,13 +6,15 @@ export 'simple_page_locale.lg.dart';
 
 /// Parting .freezed.dart and .g.dart files are required by freezed and json_serializable
 part 'simple_page_locale.freezed.dart';
+
 part 'simple_page_locale.g.dart';
 
 /// LocaleGen annotation is required for locale_gen to work
 /// freezed annotation is required for freezed to work
 @LocaleGen()
 @freezed
-class SimplePageLocale with _$SimplePageLocale {
+abstract class SimplePageLocale with _$SimplePageLocale {
+  @JsonSerializable(converters: [])
   const factory SimplePageLocale({
     required String title,
     required String subtitle,