+++ locale_generator/example/lib/simple_page_locale/simple_page_locale.freezed.dart
@@ -1,3 +1,4 @@
+// dart format width=80
 // coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
 // ignore_for_file: type=lint
@@ -9,96 +10,77 @@ part of 'simple_page_locale.dart';
 // FreezedGenerator
 // **************************************************************************
 
+// dart format off
 T _$identity<T>(T value) => value;
-
-final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
-
-SimplePageLocale _$SimplePageLocaleFromJson(Map<String, dynamic> json) {
-  return _SimplePageLocale.fromJson(json);
-}
 
 /// @nodoc
 mixin _$SimplePageLocale {
-  String get title => throw _privateConstructorUsedError;
-  String get subtitle => throw _privateConstructorUsedError;
+  String get title;
+  String get subtitle;
 
-  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
-  @JsonKey(ignore: true)
-  $SimplePageLocaleCopyWith<SimplePageLocale> get copyWith =>
-      throw _privateConstructorUsedError;
-}
-
-/// @nodoc
-abstract class $SimplePageLocaleCopyWith<$Res> {
-  factory $SimplePageLocaleCopyWith(
-          SimplePageLocale value, $Res Function(SimplePageLocale) then) =
-      _$SimplePageLocaleCopyWithImpl<$Res, SimplePageLocale>;
-  @useResult
-  $Res call({String title, String subtitle});
-}
-
-/// @nodoc
-class _$SimplePageLocaleCopyWithImpl<$Res, $Val extends SimplePageLocale>
-    implements $SimplePageLocaleCopyWith<$Res> {
-  _$SimplePageLocaleCopyWithImpl(this._value, this._then);
-
-  // ignore: unused_field
-  final $Val _value;
-  // ignore: unused_field
-  final $Res Function($Val) _then;
-
+  /// Create a copy of SimplePageLocale
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
   @pragma('vm:prefer-inline')
+  $SimplePageLocaleCopyWith<SimplePageLocale> get copyWith =>
+      _$SimplePageLocaleCopyWithImpl<SimplePageLocale>(
+          this as SimplePageLocale, _$identity);
+
+  /// Serializes this SimplePageLocale to a JSON map.
+  Map<String, dynamic> toJson();
+
   @override
-  $Res call({
-    Object? title = null,
-    Object? subtitle = null,
-  }) {
-    return _then(_value.copyWith(
-      title: null == title
-          ? _value.title
-          : title // ignore: cast_nullable_to_non_nullable
-              as String,
-      subtitle: null == subtitle
-          ? _value.subtitle
-          : subtitle // ignore: cast_nullable_to_non_nullable
-              as String,
-    ) as $Val);
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is SimplePageLocale &&
+            (identical(other.title, title) || other.title == title) &&
+            (identical(other.subtitle, subtitle) ||
+                other.subtitle == subtitle));
+  }
+
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  @override
+  int get hashCode => Object.hash(runtimeType, title, subtitle);
+
+  @override
+  String toString() {
+    return 'SimplePageLocale(title: $title, subtitle: $subtitle)';
   }
 }
 
 /// @nodoc
-abstract class _$$SimplePageLocaleImplCopyWith<$Res>
-    implements $SimplePageLocaleCopyWith<$Res> {
-  factory _$$SimplePageLocaleImplCopyWith(_$SimplePageLocaleImpl value,
-          $Res Function(_$SimplePageLocaleImpl) then) =
-      __$$SimplePageLocaleImplCopyWithImpl<$Res>;
-  @override
+abstract mixin class $SimplePageLocaleCopyWith<$Res> {
+  factory $SimplePageLocaleCopyWith(
+          SimplePageLocale value, $Res Function(SimplePageLocale) _then) =
+      _$SimplePageLocaleCopyWithImpl;
   @useResult
   $Res call({String title, String subtitle});
 }
 
 /// @nodoc
-class __$$SimplePageLocaleImplCopyWithImpl<$Res>
-    extends _$SimplePageLocaleCopyWithImpl<$Res, _$SimplePageLocaleImpl>
-    implements _$$SimplePageLocaleImplCopyWith<$Res> {
-  __$$SimplePageLocaleImplCopyWithImpl(_$SimplePageLocaleImpl _value,
-      $Res Function(_$SimplePageLocaleImpl) _then)
-      : super(_value, _then);
+class _$SimplePageLocaleCopyWithImpl<$Res>
+    implements $SimplePageLocaleCopyWith<$Res> {
+  _$SimplePageLocaleCopyWithImpl(this._self, this._then);
 
+  final SimplePageLocale _self;
+  final $Res Function(SimplePageLocale) _then;
+
+  /// Create a copy of SimplePageLocale
+  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
     Object? title = null,
     Object? subtitle = null,
   }) {
-    return _then(_$SimplePageLocaleImpl(
+    return _then(_self.copyWith(
       title: null == title
-          ? _value.title
+          ? _self.title
           : title // ignore: cast_nullable_to_non_nullable
               as String,
       subtitle: null == subtitle
-          ? _value.subtitle
+          ? _self.subtitle
           : subtitle // ignore: cast_nullable_to_non_nullable
               as String,
     ));
@@ -106,66 +88,91 @@ class __$$SimplePageLocaleImplCopyWithImpl<$Res>
 }
 
 /// @nodoc
-@JsonSerializable()
-class _$SimplePageLocaleImpl implements _SimplePageLocale {
-  const _$SimplePageLocaleImpl({required this.title, required this.subtitle});
 
-  factory _$SimplePageLocaleImpl.fromJson(Map<String, dynamic> json) =>
-      _$$SimplePageLocaleImplFromJson(json);
+@JsonSerializable(converters: [])
+class _SimplePageLocale implements SimplePageLocale {
+  const _SimplePageLocale({required this.title, required this.subtitle});
+  factory _SimplePageLocale.fromJson(Map<String, dynamic> json) =>
+      _$SimplePageLocaleFromJson(json);
 
   @override
   final String title;
   @override
   final String subtitle;
 
+  /// Create a copy of SimplePageLocale
+  /// with the given fields replaced by the non-null parameter values.
   @override
-  String toString() {
-    return 'SimplePageLocale(title: $title, subtitle: $subtitle)';
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  @pragma('vm:prefer-inline')
+  _$SimplePageLocaleCopyWith<_SimplePageLocale> get copyWith =>
+      __$SimplePageLocaleCopyWithImpl<_SimplePageLocale>(this, _$identity);
+
+  @override
+  Map<String, dynamic> toJson() {
+    return _$SimplePageLocaleToJson(
+      this,
+    );
   }
 
   @override
   bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
-            other is _$SimplePageLocaleImpl &&
+            other is _SimplePageLocale &&
             (identical(other.title, title) || other.title == title) &&
             (identical(other.subtitle, subtitle) ||
                 other.subtitle == subtitle));
   }
 
-  @JsonKey(ignore: true)
+  @JsonKey(includeFromJson: false, includeToJson: false)
   @override
   int get hashCode => Object.hash(runtimeType, title, subtitle);
 
-  @JsonKey(ignore: true)
   @override
-  @pragma('vm:prefer-inline')
-  _$$SimplePageLocaleImplCopyWith<_$SimplePageLocaleImpl> get copyWith =>
-      __$$SimplePageLocaleImplCopyWithImpl<_$SimplePageLocaleImpl>(
-          this, _$identity);
-
-  @override
-  Map<String, dynamic> toJson() {
-    return _$$SimplePageLocaleImplToJson(
-      this,
-    );
+  String toString() {
+    return 'SimplePageLocale(title: $title, subtitle: $subtitle)';
   }
 }
 
-abstract class _SimplePageLocale implements SimplePageLocale {
-  const factory _SimplePageLocale(
-      {required final String title,
-      required final String subtitle}) = _$SimplePageLocaleImpl;
-
-  factory _SimplePageLocale.fromJson(Map<String, dynamic> json) =
-      _$SimplePageLocaleImpl.fromJson;
-
+/// @nodoc
+abstract mixin class _$SimplePageLocaleCopyWith<$Res>
+    implements $SimplePageLocaleCopyWith<$Res> {
+  factory _$SimplePageLocaleCopyWith(
+          _SimplePageLocale value, $Res Function(_SimplePageLocale) _then) =
+      __$SimplePageLocaleCopyWithImpl;
   @override
-  String get title;
-  @override
-  String get subtitle;
-  @override
-  @JsonKey(ignore: true)
-  _$$SimplePageLocaleImplCopyWith<_$SimplePageLocaleImpl> get copyWith =>
-      throw _privateConstructorUsedError;
+  @useResult
+  $Res call({String title, String subtitle});
 }
+
+/// @nodoc
+class __$SimplePageLocaleCopyWithImpl<$Res>
+    implements _$SimplePageLocaleCopyWith<$Res> {
+  __$SimplePageLocaleCopyWithImpl(this._self, this._then);
+
+  final _SimplePageLocale _self;
+  final $Res Function(_SimplePageLocale) _then;
+
+  /// Create a copy of SimplePageLocale
+  /// with the given fields replaced by the non-null parameter values.
+  @override
+  @pragma('vm:prefer-inline')
+  $Res call({
+    Object? title = null,
+    Object? subtitle = null,
+  }) {
+    return _then(_SimplePageLocale(
+      title: null == title
+          ? _self.title
+          : title // ignore: cast_nullable_to_non_nullable
+              as String,
+      subtitle: null == subtitle
+          ? _self.subtitle
+          : subtitle // ignore: cast_nullable_to_non_nullable
+              as String,
+    ));
+  }
+}
+
+// dart format on