+++ locale_generator/CHANGELOG.md
@@ -5,3 +5,11 @@
 ## 1.0.6
 
 - Flutter compatibility fixes
+
+## 1.0.7
+
+- Added support to ignore fields with JsonKey(includeFromJson: false)
+
+## 1.1.0
+
+- Added support to custom converters with LocaleGen(converters: ...)